double gameTime and masked transparency depth check

+++ RetroEngine.Game/Entities/Bullet.cs
@@ -156,7 +156,7 @@ namespace RetroEngine.Game.Entities
 
             mesh.Position = Position;
             mesh.Rotation = startRotation;
-            mesh.Scale = new Vector3(1f-(Time.gameTime - SpawnTime) / LifeTime);
+            mesh.Scale = new Vector3((float)(1d-(Time.gameTime - SpawnTime) / LifeTime));
         }
 
         public override void LateUpdate()
