+++ RetroEngine.Game/Entities/Weapons/Weapon.cs
@@ -29,7 +29,7 @@ namespace RetroEngine.Game.Entities.Weapons
         protected WeaponData data;
         protected PlayerCharacter player;
 
-        protected float DrawTime = 0.35f;
+        protected double DrawTime = 0.35f;
 
         protected bool Drawing = true;
 
@@ -92,7 +92,7 @@ namespace RetroEngine.Game.Entities.Weapons
         {
             base.LateUpdate();
 
-            float progress = Math.Clamp((Time.gameTime - SpawnTime)/DrawTime, 0, 1);
+            float progress = Math.Clamp((float)((Time.gameTime - SpawnTime)/DrawTime), 0, 1);
 
             DrawRotation = new Vector3((1 - progress) * 20, (1 - progress) * 20, (1 - progress) * 30);
